requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-tinymce==4.0.0
djangorestframework==3.15.1
Markdown==3.6
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
